Moved Log.h/.cpp to TheOneEngine
- Changed the includes so that they are correctly linked.
