+++ TheOneEditor/TheOneEditor.vcxproj.filters
@@ -1,9 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="Log.cpp">
-      <Filter>Utils</Filter>
-    </ClCompile>
     <ClCompile Include="Panel.cpp">
       <Filter>Panels</Filter>
     </ClCompile>
@@ -69,9 +66,6 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Log.h">
-      <Filter>Utils</Filter>
-    </ClInclude>
     <ClInclude Include="Panel.h">
       <Filter>Panels</Filter>
     </ClInclude>
